Rename 'Formatter' as 'Recorder', add method 'NeedTrajectory()'
