+++ Meteorites.CpuSolvers/GoldSolver.h
@@ -1,6 +1,8 @@
 #pragma once
 #include "Meteorites.Core/Defs.h"
+
 #include "Meteorites.Core/BasicSolver.h"
+
 
 // Uses one-step, two-step or three-step Adams method to solve ordinary differential equations
 // May be used to verify more complex solvers with some performance optimizations
@@ -12,5 +14,5 @@ class GoldSolver : public BasicSolver
     // ISolver method
     virtual void Solve(const VirtualMeteoroid &problem,
                        const IFunctional &functional,
-                       IResultFormatter &results) override final;
+                       ISimulationRecorder &results) override final;
 };
